Prefactor for adding callout for swipe events.
