+++ apps/style/common.scss
@@ -1713,4 +1713,4 @@ a.WireframeButtons_active, div.WireframeButtons_active {
 }
 
 @import "RotateContainer";
-@import "HideToolbarHelper";
+@import "Callout";
